custom button section silindi button group olarak düzenlendi
